+++ index.js
@@ -15,9 +15,14 @@ const volumeCvt = document.getElementById("volume-cvt")
 const massCvt = document.getElementById("mass-cvt")
 
 
+let measure = {  
+    //let length,
+    //volume,
+    //mass,
+    //function convert(){}
+}
 
 convertBtn.addEventListener("click", function(){
-    
     
     console.log("clicked")
     
@@ -25,33 +30,57 @@ convertBtn.addEventListener("click", function(){
     
     console.log(convertToFeet(measureElt.value))
     
-    displayConvertion(measureElt.value,lengthCvt)
+    displayConversion(measureElt.value,lengthCvt)
+    
+    displayConversion(measureElt.value,volumeCvt)
     
 }
 
-)
+) 
 
+// Meters to Feet
 function convertToFeet(lengthMeter){//1 meter = 3.281 feet
     
     const lengthFeet = (Number(lengthMeter) / 3.281)
     
-    console.log(`${lengthMeter} meter > ${lengthFeet} feet`)
+    console.log(`${lengthMeter} meters = ${lengthFeet} feet`)
     
     return roundedNumber(lengthFeet)
 }
 
-
+// Feet to Meters
 function convertToMeter(lengthFeet){//1 meter = 3.281 feet
     
     const lengthMeter = (Number(lengthFeet) * 3.281)
     
-    console.log(`${lengthFeet} feet > ${lengthMeter} meters`)
+    console.log(`${lengthFeet} feet = ${lengthMeter} meters`)
     
     return roundedNumber(lengthMeter)
 }
 
 
-function displayConvertion(measure, element){
+// Liters to Gallons
+function convertToGallons(volumeLiter){//1 liter = 0.264 gallon
+    
+    const volumeGallon = (Number(volumeLiter) / 0.264)
+    
+    console.log(`${volumeLiter} liters = ${volumeGallon} gallons`)
+    
+    return roundedNumber(volumeLiter)
+}
+
+// Gallons to Liters
+function convertToLiter(volumeGallon){//1 liter = 0.264 gallon
+    
+    const volumeLiter = (Number(volumeGallon) * 0.264)
+    
+    console.log(`${volumeGallon} gallons = ${volumeLiter} liters`)
+    
+    return roundedNumber(volumeGallon)
+}
+
+
+function displayConversion(measure, element){
     
     
     element.innerHTML =`