+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 ## [Unreleased]
 - CSS Designing features upcoming
+- Convert the mass from Kilograms to Pounds
+- Convert the mass from Pounds to Kilograms
+
+## [0.2.0] - 2025-11-28
+### Added
+- Implemented a function to convert Liters values to Gallons
+- Implemented a function to convert Gallons values to Liters
 
 ## [0.1.0] - 2025-11-27
 ### Added
 - Implemented a function to round numbers down to 3 decimal places
-- Implemented a function to convert values to meters
+- Implemented a function to convert Feet values to Meters
 
 ### Changed
 - Improved function to display converted measures